pyproject.toml
[project]
name = "crewai_example"
version = "0.1.0"
description = "crewai-example using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.100.0,<1.0.0",
    "sandboxai-client",
]

[project.scripts]
crewai_example = "crewai_example.main:run"
run_crew = "crewai_example.main:run"
train = "crewai_example.main:train"
replay = "crewai_example.main:replay"
test = "crewai_example.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
